+++ CHANGES.md
@@ -8,7 +8,8 @@ Improvements:
  - Persist active tab between sessions (#503)
  - Do not upload file too big for the homeserver (#587)
  - Handle read markers (#84)
- - Attachments: start using system pickers
+ - Attachments: start using system pickers (#52)
+ - Attachments: start handling incoming share (#58)
 
 Other changes:
  - Accessibility improvements to read receipts in the room timeline and reactions emoji chooser